pyproject.toml
[tool.poetry]
name = "kitchenview"
version = "0.1.0"
description = ""
authors = ["Kirill Nikolaevskiy <kirill.nikolaevskiy@rt.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.1"
aiohttp = "^3.9.5"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
asyncio = "^3.4.3"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pydantic = "2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
